requirements.txt
PyYaml
grpc
scapy >= 2.4.2
grpcio==1.20
requests>=2.19.1
urllib3>=1.23
logging
psutil
anytree>=2.6.0
pydevd
gevent==1.1.0
python-dateutil
mock
